+++ src/Api/Enrollments/Enrollment.php
@@ -1047,4 +1047,15 @@ class Enrollment extends AbstractBaseApi
             throw new CanvasApiException("Could not load section with ID {$this->sectionId}: " . $e->getMessage());
         }
     }
+
+    /**
+     * Get the API endpoint for this resource
+     * @return string
+     * @throws CanvasApiException
+     */
+    protected static function getEndpoint(): string
+    {
+        self::checkCourse();
+        return sprintf('courses/%d/enrollments', self::$course->getId());
+    }
 }